Fixed bug 294
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code@3677 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ Palladio.SensorFactory/trunk/de.uka.ipd.sdq.sensorframework/src-man/de/uka/ipd/sdq/sensorfactory/dao/db4o/DB4ODAOFactory.java
@@ -128,7 +128,7 @@ public class DB4ODAOFactory implements IDAOFactory {
 		return this.id;
 	}
 
-	public void setID(int i) {
+	public void setID(long i) {
 		this.id = i;
 	}
