+++ Palladio.SensorFactory/trunk/de.uka.ipd.sdq.sensorframework/src-man/de/uka/ipd/sdq/sensorfactory/dao/file/FileDAOFactory.java
@@ -164,7 +164,7 @@ public class FileDAOFactory implements IDAOFactory {
 		return null;
 	}
 
-	public void setID(int i) {
+	public void setID(long i) {
 		factoryID = i;
 	}
