+++ Palladio.SensorFactory/trunk/de.uka.ipd.sdq.sensorframework/src-man/de/uka/ipd/sdq/sensorfactory/dao/memory/MemoryDAOFactory.java
@@ -91,7 +91,7 @@ public class MemoryDAOFactory implements IDAOFactory {
 		return this.id;
 	}
 
-	public void setID(int i) {
+	public void setID(long i) {
 		this.id = i;
 	}
